refactor: Tema movido para src/theme
