+++ src/theme/components/Footer.js
