+++ src/theme/components/Segment.js
