+++ src/theme/components/Separator.js
