+++ src/theme/components/SwipeRow.js
