+++ src/theme/components/Thumbnail.js
